+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 # HEAD
+- Functions to decide which type of render use
+- Functions to get the color from a pigment and evaluete it thanks to the BRDF
+- New structs for rendering (`OnOffRenderer`, `FlatRRenderer`) and tests
+- New struct for implement the material of a Shape (`Material`) and tests
+- New structs for BRDF implementtion (`BRDF`, `DiffuseBRDF`) and tests
+- New structs for pigment implementation (`Pigment`, `UniformPigment`, `CheckeredPigment`, `ImagePigment`) and tests
 - Modified operation between `Transformation` and `Normal`, `Vec` and `Point` for time optimization
 - Added test for World, modified test for Sphere
 - Added feature: now can create an animation